LayoutTests/imported/w3c/web-platform-tests/cors/cors-safelisted-request-header.any.worker-expected.txt

Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=95a59327-e440-437a-a9c6-1c2135b5f8ac
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=409ac348-467e-425f-af9a-b3cf23411c7e
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=8aff817f-b2e9-4a90-b691-50b40b061e90
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=301859c5-6339-41da-bbe4-f5b6851ca6f5
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=fefc5ca0-ddc7-49d1-bb37-9b169927f6f3
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=af67419d-62fb-4a28-9c40-b346bca62f46
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=7c428628-9fc6-43fe-93ab-4ea27d6718ab
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=77c92f10-e552-46d8-9efd-b98ab2746646
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=93432db8-732f-416f-88cc-2d33622695b2
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=d4ec8da6-1539-40de-be57-ec0765588ddd
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=7a224dfc-4247-451f-ac3b-afb2d0aec163
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=73361270-95fc-4a9f-883d-0ec41226f03d
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=c0b09995-56ae-433d-9713-014e47df2dcb
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=6f6ead2c-4661-46ca-9952-66fa54e3c886
Blocked access to external URL http://www1.localhost:8800/cors/resources/preflight.py?token=07c99b69-29d0-45ec-a4b9-d375d645cc7c

FAIL Preflight for {"content-type":"text /plain"} assert_equals: expected "1" but got "0"
FAIL Preflight for {"content-type":"text\t/\tplain"} assert_equals: expected "1" but got "0"
FAIL No preflight for {"content-type":"text/plain;"} promise_test: Unhandled rejection with value: object "TypeError: Type error"
FAIL No preflight for {"content-type":"text/plain;garbage"} promise_test: Unhandled rejection with value: object "TypeError: Type error"
FAIL Preflight for {"content-type":"text/plain;garbage\u0001\u0002"} assert_equals: expected "1" but got "0"
FAIL Preflight for {"content-type":"text/plain,"} assert_equals: expected "1" but got "0"
FAIL Preflight for {"content-type":",text/plain"} assert_equals: expected "1" but got "0"
FAIL Preflight for {"content-type":"text/plain,text/plain"} assert_equals: expected "1" but got "0"
FAIL Preflight for {"content-type":"text/plain,x/x"} assert_equals: expected "1" but got "0"
FAIL Preflight for {"content-type":"text/plain\u000b"} assert_equals: expected "1" but got "0"
FAIL Preflight for {"content-type":"text/plain\f"} assert_equals: expected "1" but got "0"
FAIL Preflight for {"content-type":"application/www-form-urlencoded"} assert_equals: expected "1" but got "0"
FAIL Preflight for {"content-type":"application/x-www-form-urlencoded;"} assert_equals: expected "1" but got "0"
FAIL No preflight for {"content-type":"multipart/form-data"} promise_test: Unhandled rejection with value: object "TypeError: Type error"
FAIL Preflight for {"content-type":"multipart/form-data;\""} assert_equals: expected "1" but got "0"
